+++ tests/Functional/TestApplication/Entity/Post.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sofascore\PurgatoryBundle\Tests\Functional\TestApplication\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Annotation\Groups as AnnotationGroups;
+use Symfony\Component\Serializer\Attribute\Groups;
+
+#[ORM\Entity]
+class Post
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    public int $id;
+
+    #[ORM\Column]
+    #[Groups('common')]
+    #[AnnotationGroups('common')]
+    public string $title;
+
+    #[ORM\Column]
+    #[Groups('extra')]
+    #[AnnotationGroups('extra')]
+    public string $text;
+
+    #[ORM\Column]
+    public int $views = 0;
+}